Fixed typo in runtime lib
Ping @MelodyLucid: you have to update the version of the miniboxing plugin to 0.3-SNAPSHOT
and the scala version to 2.11.1 to use the new plugin. See this link:
https://github.com/miniboxing/miniboxing-example/commit/a50b3ac2058fa2036b75cb64adee413f9f500e50#diff-1

+++ components/runtime/src/miniboxing/runtime/MiniboxDispatchDouble.java
@@ -51,7 +51,7 @@ public class MiniboxDispatchDouble {
   }
 
   // non-overloaded:
-  public final static boolean mboxed_eqeq_notag(double x, long y) {
+  public final static boolean mboxed_eqeq_notag(double x, double y) {
     return mboxed_eqeq(x, y);
   }
